# Use a codeunit type-assert that is easier for inference to elide