--- a/base/regex.jl
+++ b/base/regex.jl
@@ -428,7 +428,7 @@ struct SubstitutionString{T<:AbstractString} <: AbstractString
 end
 
 ncodeunits(s::SubstitutionString) = ncodeunits(s.string)::Int
-codeunit(s::SubstitutionString) = codeunit(s.string)::Type{<:Union{UInt8, UInt16, UInt32}}
+codeunit(s::SubstitutionString) = codeunit(s.string)::CodeunitType
 codeunit(s::SubstitutionString, i::Integer) = codeunit(s.string, i)::Union{UInt8, UInt16, UInt32}
 isvalid(s::SubstitutionString, i::Integer) = isvalid(s.string, i)::Bool
 iterate(s::SubstitutionString, i::Integer...) = iterate(s.string, i...)::Union{Nothing,Tuple{AbstractChar,Int}}