Circumvents #37135