@omus 2020-09-04
I was working on a performance improvement for `==(::AbstractString, ::AbstractString)` but I discovered this PR addresses the performance issue I was noticing (mostly allocations):

Before this PR (6aee988223):

```julia
julia> str = "A" ^ 1000;

julia> @btime invoke(==, Tuple{AbstractString, AbstractString}, $"ABC", $"ABC")
  4.357 ns (0 allocations: 0 bytes)
true

julia> @btime $(SubString("ABC")) == $"ABCD"
  29.574 ns (2 allocations: 64 bytes)
false

julia> @btime $(SubString("ABC")) == $"ABC"
  32.492 ns (2 allocations: 64 bytes)
true

julia> @btime $(SubString("ABC")) == $"DEF"
  31.985 ns (2 allocations: 64 bytes)
false

julia> @btime $(SubString(str)) == $str
  65.086 ns (2 allocations: 64 bytes)
true
```

With this PR (dbf5e005d9):

```julia
julia> str = "A" ^ 1000;

julia> @btime invoke(==, Tuple{AbstractString, AbstractString}, $"ABC", $"ABC")
  3.907 ns (0 allocations: 0 bytes)
true

julia> @btime $(SubString("ABC")) == $"ABCD"
  0.033 ns (0 allocations: 0 bytes)
false

julia> @btime $(SubString("ABC")) == $"ABC"
  4.003 ns (0 allocations: 0 bytes)
true

julia> @btime $(SubString("ABC")) == $"DEF"
  3.282 ns (0 allocations: 0 bytes)
false

julia> @btime $(SubString(str)) == $str
  38.978 ns (0 allocations: 0 bytes)
true
```

I'll be adding some benchmarks for this.