@KristofferC 2020-09-05
Is this v1.5? Because AFAIU this just fixed an earlier regression on nightly. But good to have benchmarks anyway.