@omus 2020-09-08
I noticed similar slow performance on Julia 1.3.1, 1.4.2, and 1.5.1. All of these versions were substantially worse with the final test:
```julia
julia> @btime $(SubString(str)) == $str
  4.043 μs (2 allocations: 112 bytes)
true
```