@KristofferC 2020-09-08
Seems more likely that the improvement here was because of https://github.com/JuliaLang/julia/pull/35973