@omus 2020-09-08
You are correct that https://github.com/JuliaLang/julia/pull/35973 is the main driver behind the performance improvement. I'll update my benchmarks to reflect this.