@omus 2020-09-08
I've also opened https://github.com/JuliaLang/julia/pull/37467 which improves performance of the fallback `==(::AbstractString, ::AbstractString)` call.